Move the WebDataStreamNotAvailableNotification component into the Analytics module.
